requirements.txt
einops==0.6.1
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.8.0.74
pandas==2.0.3
pillow==10.0.0
plotnine==0.12.4
pyyaml==6.0.1
scikit_image==0.21.0
scikit_learn==1.3.0
scipy==1.10.1
tensorboardX==2.6.1
torch==1.8.0+cu111
torchvision==0.9.0
tqdm==4.65.0